requirements.txt
openai-whisper
spacy
phonenumbers
Flask
